Move powerOfTen and absolute to dedicated modules
- Move BIGINT_POWERS_OF_TEN and powerOfTen to src/functions/utils/index.ts
- Move absolute function from rounding.ts to arithmetic/operations.ts
- Rename absoluteValue to absolute
- Update imports accordingly
