+++ src/functions/arithmetic/power.ts
