+++ src/functions/numeric/rounding.ts
@@ -1,40 +1,5 @@
 import type { FPContext, RoundingMode } from "../../FixedPrecision";
-
-const BIGINT_POWERS_OF_TEN: bigint[] = [
-  1n,
-  10n,
-  100n,
-  1_000n,
-  10_000n,
-  100_000n,
-  1_000_000n,
-  10_000_000n,
-  100_000_000n,
-  1_000_000_000n,
-  10_000_000_000n,
-  100_000_000_000n,
-  1_000_000_000_000n,
-  10_000_000_000_000n,
-  100_000_000_000_000n,
-  1_000_000_000_000_000n,
-  10_000_000_000_000_000n,
-  100_000_000_000_000_000n,
-  1_000_000_000_000_000_000n,
-  10_000_000_000_000_000_000n,
-  100_000_000_000_000_000_000n,
-] as const;
-
-function powerOfTen(exponent: number): bigint {
-  const value = BIGINT_POWERS_OF_TEN[exponent];
-  if (value === undefined) {
-    throw new Error("Decimal places must be an integer between 0 and 20");
-  }
-  return value;
-}
-
-export function absoluteValue(value: bigint): bigint {
-  return value < 0n ? -value : value;
-}
+import { powerOfTen } from "../utils";
 
 export function roundToScaleValue(
   value: bigint,
